1-3 미션
createUserFromUrl 메서드 AddressHandler -> User 이동
