+++ src/main/java/webserver/AddressHandler.java
@@ -6,6 +6,7 @@ import java.io.DataOutputStream;
 import java.io.IOException;
 
 import static db.Database.addUser;
+import static model.User.createUserFromUrl;
 
 public class AddressHandler {
 
@@ -19,19 +20,6 @@ public class AddressHandler {
         return filePath;
     }
 
-    public static User createUserFromUrl(String urlPath) {
-        int parameterIndex = urlPath.indexOf("?");
-        if (parameterIndex != -1) {
-            String[] userInfo = urlPath.substring(parameterIndex + 1).split("&");
-            return new User(
-                    userInfo[0].substring(userInfo[0].indexOf("=") + 1),
-                    userInfo[1].substring(userInfo[1].indexOf("=") + 1),
-                    userInfo[2].substring(userInfo[2].indexOf("=") + 1),
-                    userInfo[3].substring(userInfo[3].indexOf("=") + 1)
-            );
-        }
-        return null;
-    }
 
     public static void handleUserCreation(String urlPath, DataOutputStream dos) throws IOException {
         User user = createUserFromUrl(urlPath);
